chore: module based file structure: event types (#13404)
